+++ src/shared/hooks/index.ts
@@ -1,17 +1,6 @@
-import useLaptop from "./use-laptop";
-import useGsap from "./use-gsap";
-import useCursor from "./use-cursor";
 import useDay from "./use-day";
 import useCheckOS from "./use-check-os";
 import useWindows from "./use-windows";
 import useApplication from "./use-application";
 
-export {
-  useLaptop,
-  useGsap,
-  useCursor,
-  useDay,
-  useCheckOS,
-  useWindows,
-  useApplication,
-};
+export { useDay, useCheckOS, useWindows, useApplication };
