--- src/shared/resources/index.ts
@@ -1,1 +1,0 @@
-export { default as LightController } from "./light-controller";
