--- src/widgets/models/index.ts
@@ -1,4 +1,0 @@
-export { default as Macbook } from "./macbook";
-export { default as Coffee } from "./coffee";
-export { default as CoffeeShop } from "./coffee-shop";
-export { default as ObjectRender } from "./object-render";
