+++ src/widgets/render-models/lib/use-laptop.ts
