+++ src/widgets/render-models/ui/starbucks-coffee.tsx
@@ -2,7 +2,7 @@ import { useGLTF } from "@react-three/drei";
 
 import { DreiGLTF } from "@/shared/types/three/gltf";
 
-const CoffeeShop = () => {
+const StarBucksCoffee = () => {
   const dir = import.meta.env.DEV ? "dev-models" : "models";
   const { scene } = useGLTF(`../${dir}/gltf/starbucks/scene.gltf`) as DreiGLTF;
 
@@ -15,4 +15,4 @@ const CoffeeShop = () => {
   );
 };
 
-export default CoffeeShop;
+export default StarBucksCoffee;
